testing a way to authenticate
